requirements.txt
asgiref==3.3.1
Django==3.1.5
django-debug-toolbar==3.2
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2020.5
sqlparse==0.4.1
